commentInput: Ensure react-md-editor is not part of the main chunk
This reduces the main chunk size from 3.65 MB to 2.56 MB.
Something about the commit "Use CRACO to workaround CRA #11770" caused
several large components to be added to the main.*.js bundle.

All suspense elements are *roughly* the same size as the expected text
editor box, plus or minus a row.

Signed-off-by: Taylor Smock <tsmock@meta.com>