Rapid: Use system interface for setting up Rapid
Also, import Rapid from CDN instead of bundling it with the TM.
We still have to bundle the Rapid CSS file -- initial try on
importing the CSS file from the CDN failed. If that is fixed,
we could use a semver query for the CDN version, and get minor
Rapid updates for "free".

Signed-off-by: Taylor Smock <tsmock@meta.com>